% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actMap}
\alias{actMap}
\title{actMap}
\usage{
actMap(subject, session, time = c(0, 10), file_path = getwd(),
  file_ext = "txt")
}
\arguments{
\item{subject}{Subject ID corresponding to file with activity time series.}

\item{session}{Session number corresponding to folder where file is located.}

\item{file_path}{Optional specification of file path where session folders are located.
Defaults to current working directory.}

\item{time}{Optional specification of a time vector specifying range in minutes to plot. Default
is the entire 10 minutes: c(0,10)}
}
\value{
plot3D object
}
\description{
Activity map: 3D plot of exploratory path taken by rodent in the activity chamber.
}
\details{
WARNING: make sure you have first run "fixData" or have otherwise interpolated missing X,Y
 coordinates. Otherwise you will get false labeling of zones for time points where rat is
 "off grid", i.e., not breaking any infrared beams, which can happen when the rat rears upright
 into a corner.
}
\author{
Jason Shumake
}

