% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRearData}
\alias{getRearData}
\title{getRearData}
\usage{
getRearData(subject, session, file_path = getwd(), file_ext = "txt")
}
\arguments{
\item{subject}{subject ID corresponding to file with activity time series.}

\item{session}{session number corresponding to folder where file is located.}

\item{file_path}{optional specification of file path where session folders are located.
Defaults to current working directory.}

\item{file_ext}{optional specification of file extension if something other than "txt".}
}
\value{
data frame containing subject, session, zone, rearing count, start time,
 duration, minimum rearing duration, maximum rearing duration, mean rearing duration, and
 rearing variability score)
}
\description{
Computes rearing counts, duration, and summary stats for a single subject.
}
\details{
WARNING: this function should only be run on files that are analysis ready, i.e.,
 you've already run "fixData" and "labelZones" on all files.
}
\author{
Jason Shumake
}

